@@ -1,8 +1,9 @@
 {
   "name": "eslint-plugin-bedrock",
-  "version": "1.0.26",
+  "version": "2.0.0",
   "description": "Common ESLint plugin for Bedrock projects.",
   "license": "MIT",
+  "type": "module",
   "keywords": [
     "eslint",
     "eslintplugin",
@@ -10,7 +11,7 @@
   ],
   "main": "index.js",
   "peerDependencies": {
-    "eslint": ">=8",
+    "eslint": ">=9",
     "eslint-import-resolver-webpack": "^0.13.2"
   },
   "peerDependenciesMeta": {
@@ -19,8 +20,12 @@
     }
   },
   "dependencies": {
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-react": "^7.32.2"
+    "@eslint/js": "^9.1.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^28.3.0",
+    "eslint-plugin-react": "^7.34.1"
+  },
+  "devDependencies": {
+    "eslint": "^9.1.1"
   }
 }
